// src/types/purchases.type.ts
import { Product } from './products.type'
export type PurchaseStatus=-1|1|2|3|4|5
export type PurchaseListStatus= PurchaseStatus | 0 
export interface Purchase {
  _id: string
  buy_count: number
  price: number
  price_before_discount: number
  status: number
  user: string
  product: Product
  createAt: string
  updateAt: string
}